refactor(swift-sdk): remove stale artifacts
