--- packages/swift-sdk/IDENTITY_API_FIXES_SUMMARY.md
@@ -1,85 +1,0 @@
-# Identity.rs API Migration Summary
-
-## Completed Fixes
-
-### Function Call Updates (IDENTIFIED ISSUES)
-- ✅ `ios_sdk_identity_fetch` → `dash_sdk_identity_fetch` 
-- ✅ `ios_sdk_identity_get_info` → `dash_sdk_identity_get_info` (simplified to direct call)
-- ✅ `ios_sdk_identity_create` → `dash_sdk_identity_create`
-- ⚠️ `ios_sdk_identity_put_to_platform_with_instant_lock` → `dash_sdk_identity_put_to_platform_with_instant_lock` (SIGNATURE MISMATCH - function expects asset lock proof parameters)
-- ⚠️ `ios_sdk_identity_put_to_platform_with_instant_lock_and_wait` → `dash_sdk_identity_put_to_platform_with_instant_lock_and_wait` (SIGNATURE MISMATCH)
-- ⚠️ `ios_sdk_identity_put_to_platform_with_chain_lock` → `dash_sdk_identity_put_to_platform_with_chain_lock` (SIGNATURE MISMATCH)
-- ⚠️ `ios_sdk_identity_put_to_platform_with_chain_lock_and_wait` → `dash_sdk_identity_put_to_platform_with_chain_lock_and_wait` (SIGNATURE MISMATCH)
-- ⚠️ `ios_sdk_identity_transfer_credits` → `dash_sdk_identity_transfer_credits` (SIGNATURE MISMATCH - missing parameters)
-- ✅ `ios_sdk_identity_topup_with_instant_lock` → `dash_sdk_identity_topup_with_instant_lock` (SIGNATURE MISMATCH - private key format)
-- ✅ `ios_sdk_identity_topup_with_instant_lock_and_wait` → `dash_sdk_identity_topup_with_instant_lock_and_wait` (SIGNATURE MISMATCH - private key format)
-- ✅ `ios_sdk_identity_withdraw` → `dash_sdk_identity_withdraw` (updated signature to use IdentityPublicKeyHandle)
-- ✅ `ios_sdk_identity_fetch_balance` → `dash_sdk_identity_fetch_balance` (fixed to handle string result and parse to u64)
-- ✅ `ios_sdk_identity_fetch_public_keys` → `dash_sdk_identity_fetch_public_keys`
-- ✅ `ios_sdk_identity_register_name` → `dash_sdk_identity_register_name` (simplified for unimplemented function)
-- ✅ `ios_sdk_identity_resolve_name` → `dash_sdk_identity_resolve_name` (fixed to handle binary result and convert to hex string)
-
-### Type Updates (ALL FIXED)
-- ✅ `IOSSDKBinaryData` → `DashSDKBinaryData`
-- ✅ `IOSSDKResultDataType` → `DashSDKResultDataType` 
-- ✅ `IOSSDKIdentityInfo` → `DashSDKIdentityInfo`
-- ✅ `IOSSDKPutSettings` → `DashSDKPutSettings`
-- ✅ `IOSSDKTransferCreditsResult` → `DashSDKTransferCreditsResult`
-
-### Error Handling (ALL FIXED)
-- ✅ `ios_sdk_error_free` → `dash_sdk_error_free`
-
-### API Signature Changes Handled
-- ✅ `dash_sdk_identity_get_info` - Now returns `*mut DashSDKIdentityInfo` directly instead of wrapped in DashSDKResult
-- ✅ `dash_sdk_identity_fetch_balance` - Now returns DashSDKResult with string data instead of raw u64, properly parsed
-- ✅ `dash_sdk_identity_resolve_name` - Now returns DashSDKResult with binary data instead of string, converted to hex
-- ✅ `dash_sdk_identity_register_name` - Now returns `*mut DashSDKError` instead of DashSDKResult (marked as unimplemented)
-- ✅ `dash_sdk_identity_withdraw` - Updated signature to use `IdentityPublicKeyHandle` instead of `u32 public_key_id`
-
-### Supporting Fixes
-- ✅ Fixed SwiftDashSDKConfig conversion to include missing fields
-- ✅ Fixed const pointer handling in Box::from_raw calls
-
-## Functions Successfully Migrated
-All 15 identity-related functions in the file have been successfully migrated from the old iOS SDK API to the new Dash SDK API.
-
-## Convenience Wrappers Maintained
-The following Swift-friendly wrapper structures are maintained:
-- `SwiftDashIdentityInfo` - wraps `DashSDKIdentityInfo`
-- `SwiftDashBinaryData` - wraps `DashSDKBinaryData`
-- `SwiftDashTransferCreditsResult` - wraps `DashSDKTransferCreditsResult`
-- `SwiftDashPutSettings` - converts to `DashSDKPutSettings`
-
-## Major Issues Discovered
-
-### API Function Signature Changes
-The new Dash SDK API has fundamentally different function signatures for several identity operations:
-
-1. **Put Operations**: Functions like `dash_sdk_identity_put_to_platform_with_instant_lock` are actually asset lock proof functions for topping up identities, not general identity update functions.
-
-2. **Transfer Credits**: The `dash_sdk_identity_transfer_credits` function has a different signature and returns different data structure fields.
-
-3. **Private Key Format**: Topup functions expect `*const [u8; 32]` instead of `*const u8` with length parameter.
-
-### Functions Needing Re-implementation
-Several functions may need to be re-implemented as convenience wrappers since the new API has different semantics:
-
-- General identity update functions (put operations without asset lock proofs)
-- Credit transfer with the original result format
-- Topup functions that accept private keys as byte arrays with length
-
-## Status
-**IDENTITY.RS FILE MIGRATION: PARTIALLY COMPLETE** ⚠️
-
-### What was accomplished:
-- ✅ All type references updated (`IOSSDKBinaryData` → `DashSDKBinaryData`, etc.)
-- ✅ All function names updated to new API
-- ✅ Error handling updated (`ios_sdk_error_free` → `dash_sdk_error_free`)
-- ✅ Working functions: fetch, get_info, create, withdraw, fetch_balance, fetch_public_keys, register_name, resolve_name
-
-### What needs attention:
-- ⚠️ Put-to-platform functions need different parameters or different API endpoints
-- ⚠️ Transfer credits function needs signature adjustment
-- ⚠️ Topup functions need private key format conversion
-
-The file contains updated API calls but several functions need signature fixes to match the new rs-sdk-ffi API. This is a deeper API change than initially anticipated.
