--- packages/swift-sdk/IMPLEMENTATION_SUMMARY.md
@@ -1,100 +1,0 @@
-# Swift SDK Implementation Summary
-
-## Overview
-This document summarizes the implementation of Swift bindings for the Dash Platform SDK, built on top of the rs-sdk-ffi crate.
-
-## Implemented Features
-
-### 1. SDK Core Functions
-- ✅ `swift_dash_sdk_create` - Create SDK instance
-- ✅ `swift_dash_sdk_destroy` - Destroy SDK instance  
-- ✅ `swift_dash_sdk_get_network` - Get configured network
-- ✅ `swift_dash_sdk_get_version` - Get SDK version
-- ✅ `swift_dash_sdk_init` - Initialize the SDK
-- ✅ Config helpers for mainnet, testnet, and local networks
-
-### 2. Data Contract Operations
-- ✅ `swift_dash_data_contract_fetch` - Fetch data contract by ID
-- ✅ `swift_dash_data_contract_get_history` - Get data contract history
-- ✅ `swift_dash_data_contract_create` - Create new data contract
-- ⚠️  `swift_dash_data_contract_put_to_platform` - Marked as not implemented (FFI not exported)
-- ⚠️  `swift_dash_data_contract_put_to_platform_and_wait` - Marked as not implemented (FFI not exported)
-- ✅ `swift_dash_data_contract_destroy` - Free data contract handle
-- ✅ `swift_dash_data_contract_info_free` - Free data contract info
-
-### 3. Document Operations
-- ✅ `swift_dash_document_fetch` - Fetch document by ID
-- ✅ `swift_dash_document_search` - Search for documents
-- ✅ `swift_dash_document_create` - Create new document
-- ✅ `swift_dash_document_put_to_platform` - Put document to platform
-- ✅ `swift_dash_document_put_to_platform_and_wait` - Put document and wait
-- ✅ `swift_dash_document_replace_on_platform` - Replace document
-- ✅ `swift_dash_document_replace_on_platform_and_wait` - Replace and wait
-- ✅ `swift_dash_document_delete` - Delete document
-- ✅ `swift_dash_document_delete_and_wait` - Delete and wait
-- ✅ `swift_dash_document_destroy` - Free document handle
-- ✅ `swift_dash_document_info_free` - Free document info
-
-### 4. Identity Operations  
-- ✅ `swift_dash_identity_fetch` - Fetch identity by ID
-- ✅ `swift_dash_identity_get_balance` - Get identity balance
-- ✅ `swift_dash_identity_resolve_name` - Resolve DPNS name
-- ✅ `swift_dash_identity_transfer_credits` - Transfer credits between identities
-- ✅ `swift_dash_identity_put_to_platform_with_instant_lock` - Put identity with instant lock
-- ✅ `swift_dash_identity_put_to_platform_with_instant_lock_and_wait` - Put identity and wait
-- ✅ `swift_dash_identity_create_note` - Helper note for identity creation process
-- ✅ `swift_dash_identity_destroy` - Free identity handle
-- ✅ `swift_dash_identity_info_free` - Free identity info
-- ✅ `swift_dash_transfer_credits_result_free` - Free transfer result
-
-### 5. Token Operations
-- ✅ `swift_dash_token_get_total_supply` - Get token total supply
-- ✅ `swift_dash_token_transfer` - Transfer tokens
-- ✅ `swift_dash_token_mint` - Mint new tokens
-- ✅ `swift_dash_token_burn` - Burn tokens
-- ✅ `swift_dash_token_info_free` - Free token info
-
-### 6. Signer Interface
-- ✅ `swift_dash_signer_create` - Create signer with callbacks
-- ✅ `swift_dash_signer_free` - Free signer
-- ✅ `swift_dash_signer_can_sign` - Test if signer can sign
-- ✅ `swift_dash_signer_sign` - Sign data
-
-### 7. Error Handling
-- ✅ Comprehensive error codes
-- ✅ Error conversion from FFI errors
-- ✅ Binary data handling
-- ✅ Memory management functions
-
-## Architecture
-
-The Swift SDK provides a thin wrapper around the rs-sdk-ffi functions with:
-- Proper null pointer checking
-- Type conversions between Swift and FFI types
-- Memory management helpers
-- Simplified parameter structures for Swift
-
-## Testing
-
-All rs-sdk-ffi tests have been ported to Swift, including:
-- SDK initialization and configuration tests
-- Identity operation tests (21 test cases)
-- Data contract tests (16 test cases)
-- Document operation tests (15 test cases)
-- Token operation tests (9 test cases)
-- Memory management tests (14 test cases)
-
-Total: 75+ test cases
-
-## Known Limitations
-
-1. Data contract put_to_platform functions are not available because they're not exported from rs-sdk-ffi
-2. Some complex operations require proper asset lock proofs and signers which need to be implemented by the iOS app
-3. Document and identity creation require proper state transition setup
-
-## Next Steps
-
-1. The data contract put functions need to be exported in rs-sdk-ffi
-2. Additional convenience wrappers could be added for common patterns
-3. Swift Package Manager integration could be improved
-4. Example iOS app could demonstrate usage patterns
