--- packages/swift-sdk/Sources/SwiftDashSDK/KeyWallet/BIP38.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-// BIP38 functionality is not available in the current FFI
-// The bip38_encrypt_private_key and bip38_decrypt_private_key functions
-// are not present in the unified header
-// This file is kept as a placeholder to avoid Xcode build errors
