+++ packages/swift-sdk/Sources/SwiftDashSDK/KeyWallet/README.md
@@ -10,7 +10,6 @@ The KeyWallet module provides:
 - Enhanced address pool management with ManagedAccount
 - Transaction building and signing
 - Provider key generation for masternodes (BLS and EdDSA)
-- BIP38 encryption/decryption
 - Multi-wallet management with managed account collections
 
 ## Architecture
@@ -31,8 +30,7 @@ The KeyWallet module provides:
 12. **Transaction** - Transaction building, signing, and checking
 13. **ProviderKeys** - Provider key generation for masternode operations
 14. **Address** - Address validation and type detection
-15. **BIP38** - BIP38 encryption/decryption for private keys
-16. **KeyDerivation** - Low-level key derivation utilities
+15. **KeyDerivation** - Low-level key derivation utilities
 
 ### FFI Integration
 
@@ -251,23 +249,6 @@ if let summary = accountCollection.getSummary() {
     print("- Identity accounts: Registration=\(summary.hasIdentityRegistration)")
     print("- Provider accounts: Voting=\(summary.hasProviderVotingKeys)")
 }
-```
-
-### BIP38 Encryption
-
-```swift
-// Encrypt a private key
-let encrypted = try BIP38.encrypt(
-    privateKey: "cVRnH5vFxVxWFWEXLBXLcNYFKgLiC7kDiXjHEcRFQ8gfFfqH7eQA",
-    passphrase: "mypassword",
-    network: .mainnet
-)
-
-// Decrypt
-let decrypted = try BIP38.decrypt(
-    encryptedKey: encrypted,
-    passphrase: "mypassword"
-)
 ```
 
 ## Account Types
