--- packages/swift-sdk/SwiftTests/TEST_FIX_SUMMARY.md
@@ -1,35 +1,0 @@
-# Swift Test Fixes Summary
-
-## Issues Fixed
-
-1. **Type naming mismatches**: Fixed double prefixes (SwiftDashSwiftDash) in the mock implementation
-2. **Header file synchronization**: Updated both header files to match
-3. **Enum constants**: Added Swift constants file for network types and error codes  
-4. **Function signatures**: Updated mock implementation to match the unified SDK API
-5. **Memory management functions**: Added missing free functions
-6. **SDK handle types**: Changed from `UnsafeMutablePointer<SwiftDashSDKHandle>` to `OpaquePointer`
-
-## Remaining Issues
-
-1. **Document tests**: Need to update to use contract handles instead of string IDs
-2. **Identity tests**: Need to update transfer_credits to use new API with identity/signer handles
-3. **Result vs Handle returns**: Many tests expect result structs but API returns handles
-4. **Missing functions**: Some test functions (e.g., swift_dash_document_search) are not in the API
-
-## Compilation Status
-
-The mock C implementation now compiles successfully. The Swift tests have various compilation errors due to:
-- API differences between the test expectations and the unified SDK
-- Functions that return handles instead of result structs
-- Tests trying to use old API signatures
-
-## Recommendation
-
-The tests need significant refactoring to match the new unified SDK API. The main patterns to update:
-
-1. Functions that previously returned results now return handles
-2. Transfer operations now require identity handles and signer handles
-3. Document operations require contract handles instead of contract ID strings
-4. Some operations from the old API are no longer available
-
-The mock implementation is correctly structured but the tests themselves need to be updated to match the new API.
