--- packages/swift-sdk/WALLET_IMPLEMENTATION_SUMMARY.md
@@ -1,100 +1,0 @@
-# Wallet Implementation Summary
-
-## What Was Fixed
-
-### Original Issue
-The wallet creation was failing with Core Data validation errors because all addresses were being generated with the same dummy value "yDummyAddress1234567890abcdef", violating unique constraints.
-
-### Solution Implemented
-Replaced dummy address generation with real address generation using the FFI functions from the Rust `key-wallet` crate.
-
-## Changes Made
-
-### 1. WalletFFIBridge.swift - Key Derivation
-```swift
-public func deriveKey(seed: Data, path: String, network: DashNetwork) -> DerivedKey? {
-    // Now uses real FFI functions:
-    // - dash_key_xprv_from_seed: Create master key from seed
-    // - dash_key_xprv_derive_path: Derive key at BIP32 path
-    // - dash_key_xprv_private_key: Extract private key
-    // - dash_key_xprv_to_xpub: Get extended public key
-    // - dash_key_xpub_public_key: Extract public key
-}
-```
-
-### 2. WalletFFIBridge.swift - Address Generation
-```swift
-public func addressFromPublicKey(_ publicKey: Data, network: DashNetwork) -> String? {
-    // Now uses real FFI function:
-    // - dash_key_address_from_pubkey: Generate P2PKH address from public key
-}
-```
-
-### 3. WalletFFIBridge.swift - Address Validation
-```swift
-public func validateAddress(_ address: String, network: DashNetwork) -> Bool {
-    // Now uses real FFI function:
-    // - dash_key_address_validate: Validate address for network
-}
-```
-
-## FFI Functions Used
-
-The implementation now uses the following FFI functions from `dash_sdk_ffi.h`:
-
-1. **Mnemonic Functions** (already working):
-   - `dash_key_mnemonic_generate`
-   - `dash_key_mnemonic_from_phrase`
-   - `dash_key_mnemonic_phrase`
-   - `dash_key_mnemonic_to_seed`
-   - `dash_key_mnemonic_destroy`
-
-2. **Key Derivation Functions** (now implemented):
-   - `dash_key_xprv_from_seed`
-   - `dash_key_xprv_derive_path`
-   - `dash_key_xprv_to_xpub`
-   - `dash_key_xprv_private_key`
-   - `dash_key_xpub_public_key`
-   - `dash_key_xprv_destroy`
-   - `dash_key_xpub_destroy`
-
-3. **Address Functions** (now implemented):
-   - `dash_key_address_from_pubkey`
-   - `dash_key_address_validate`
-
-## Expected Behavior
-
-When creating a wallet:
-1. A mnemonic is generated (or imported)
-2. The mnemonic is converted to a 64-byte seed
-3. Keys are derived using BIP44 paths:
-   - External addresses: `m/44'/5'/0'/0/i`
-   - Internal addresses: `m/44'/5'/0'/1/i`
-4. Real Dash addresses are generated from the public keys
-5. Each address is unique and valid for the network (testnet/mainnet)
-
-## Testing
-
-To test the implementation:
-1. Build and run the SwiftExampleApp
-2. Click "Create Wallet"
-3. Enter a wallet name and PIN
-4. Optionally import a test mnemonic
-5. Click "Create"
-6. The wallet should be created successfully with unique addresses
-
-### Test Mnemonic
-```
-abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon about
-```
-
-Expected first addresses on testnet:
-- External: `yXRQqBcJXZJNXNXKqtMopfKcJu4MdNrAsc`
-- Internal: `yNPcF7DbmBGkzYksKRXMqRZpUXBpfR2fHv`
-
-## Next Steps
-
-1. Verify wallet creation works in the simulator
-2. Test address generation with different mnemonics
-3. Implement transaction signing using the private keys
-4. Add support for other address types (CoinJoin, Identity)
